+++ packages/parser/src/lint/plugin-core/lib/matchers.ts
@@ -3,18 +3,13 @@ import { CachedWildcardMatcher } from '@terrazzo/token-tools';
 /**
  * Share one cached matcher factory for all lint plugins.
  *
- * The tricky part here is creating new wildcard matchers is expensive, so we
- * want to reuse as much as possible.  However, think about all the different
- * contexts a user may need wildcard matchers across all plugins and contexts.
- * We don’t want to end up with a memory leak, potentially keeping around
- * thousands upon thousands of these things needlessly.
- *
- * A happy medium is keeping a cached wildcard matcher, but limiting its scope
- * among common codepaths. Keeping one linter cache like this yields the
- * performance benefits we want, without being a memory concern because within
- * all the linter rules this likely won’t have hundreds upon hundreds of
- * programmatically-generated instances. It will only have as many instances as
- * the user has lint rules, give or take. For other things like plugins, we want
- * to have the cache be garbage-collected more regularly, etc.
+ * Creating matchers is CPU-intensive, however, if we made one matcher for very
+ * getTransform plugin query, we could end up with tens of thousands of
+ * matchers, all taking up space in memory, but without providing any caching
+ * benefits if a matcher is used only once. So a reasonable balance is we
+ * maintain one cache per task category, and we garbage-collect everything after
+ * it’s done. Lint tasks are likely to have frequently-occurring patterns. So
+ * we’d expect for most use cases a shared lint cache has benefits, but only
+ * so long as this doesn’t spread to other plugins and other task categories.
  */
 export const cachedLintMatcher = new CachedWildcardMatcher();