+++ packages/parser/src/parse/load.ts
@@ -9,7 +9,7 @@ import {
   replaceNode,
   traverse,
 } from '@terrazzo/json-schema-tools';
-import type { TokenNormalized, TokenNormalizedSet } from '@terrazzo/token-tools';
+import type { TokenNormalizedSet } from '@terrazzo/token-tools';
 import { toMomoa } from '../lib/momoa.js';
 import { filterResolverPaths } from '../lib/resolver-utils.js';
 import type Logger from '../logger.js';