+++ packages/parser/test/parse.test.ts
@@ -833,6 +833,18 @@ describe('Additional cases', () => {
     const { tokens } = await parse([{ filename: DEFAULT_FILENAME, src }], { config });
     expect(tokens['color.blue.06']?.$value).toEqual({ alpha: 1, components: [0, 0, 1], colorSpace: 'srgb' });
   });
+
+  it('empty sources', async () => {
+    await expect(parse([], { config: defineConfig({}, { cwd: new URL('file:///') }) })).rejects.toThrow(
+      'parser:init: Nothing to parse.',
+    );
+  });
+
+  it('invalid sources', async () => {
+    await expect(
+      parse([{ filename: 'foo' } as any], { config: defineConfig({}, { cwd: new URL('file:///') }) }),
+    ).rejects.toThrow('parser:init: Input 0: expected { src: any; filename: URL }');
+  });
 });
 
 interface Visit {