+++ packages/plugin-tailwind/src/index.ts
@@ -19,11 +19,12 @@ const DEFAULT_THEME = '.';
 
 export default function pluginTailwind(options: TailwindPluginOptions): Plugin {
   const filename = options?.filename ?? 'tailwind-theme.css';
-  const variations: Record<string, Record<string, string> | undefined> = {
-    [DEFAULT_THEME]: options.defaultPermutation,
-    ...options?.variants,
+  const variations = {
+    [DEFAULT_THEME]: { selector: '', input: options.defaultTheme ?? { tzMode: '.' } },
+    ...options?.customVariants,
   };
   let cwd: URL;
+  const msg = { group: 'plugin' as const, label: PLUGIN_NAME };
 
   return {
     name: PLUGIN_NAME,
@@ -31,31 +32,25 @@ export default function pluginTailwind(options: TailwindPluginOptions): Plugin {
     config(config, { logger }) {
       if (!config.plugins.some((p) => p.name === '@terrazzo/plugin-css')) {
         logger.error({
-          group: 'plugin',
-          label: PLUGIN_NAME,
+          ...msg,
           message:
             '@terrazzo/plugin-css missing! Please install and add to the plugins array to use the Tailwind plugin.',
         });
       }
 
       if (!options || !options.theme) {
-        logger.error({ group: 'plugin', label: PLUGIN_NAME, message: 'Missing Tailwind `theme` option.' });
+        logger.error({ ...msg, message: 'Missing Tailwind `theme` option.' });
       }
 
       if (options && 'modeVariants' in options) {
-        logger.error({
-          group: 'plugin',
-          label: PLUGIN_NAME,
-          message: 'Migrate "modeVariants" to "variants" in config (see docs)',
-        });
+        logger.error({ ...msg, message: 'Migrate "modeVariants" to "variants" in config (see docs)' });
       }
 
       // store cwd for template resolution (parent of outDir)
       cwd = new URL('./', config.outDir);
       if (options?.template && !fsSync.existsSync(new URL(options.template, cwd))) {
         logger.error({
-          group: 'plugin',
-          label: PLUGIN_NAME,
+          ...msg,
           message: `Could not locate template ${fileURLToPath(new URL(options.template, cwd))}. Does the file exist?`,
         });
       }
@@ -63,19 +58,16 @@ export default function pluginTailwind(options: TailwindPluginOptions): Plugin {
     async transform({ getTransforms, setTransform, context: { logger } }) {
       const flatTheme = flattenThemeObj(options.theme);
 
-      for (const input of Object.values(variations)) {
+      for (const { input } of Object.values(variations)) {
+        const query = getTokenQuery(input);
         // Note: it’s important to remember that under-the-hood, getting/setting modes is NOT the same as having an { tzMode: value } input.
         // The former allows glob searching across all modes, the latter does not. Especially in the Tailwind plugin, confusing the two
         // will result in many dropped tokens.
-        //
-        // This is all behavior that will go away in 3.0, but for 2.0 this needs to be backwards-compatible with 0.x, hence the complexity.
-        const transformContext = input && isLegacyModes(input) ? { mode: input.tzMode } : { input };
-
         for (const { path, value } of flatTheme) {
-          const variantTokens = getTransforms({ format: FORMAT_CSS, id: value, ...transformContext });
+          const variantTokens = getTransforms({ ...query, format: FORMAT_CSS, id: value });
           // Warn the user if they are trying to generate an empty Tailwind variant
           if (!variantTokens.length) {
-            logger.warn({ group: 'plugin', label: PLUGIN_NAME, message: `${value} matched 0 tokens` });
+            logger.warn({ ...msg, message: `${value} matched 0 tokens` });
           }
 
           for (const token of variantTokens) {
@@ -85,10 +77,10 @@ export default function pluginTailwind(options: TailwindPluginOptions): Plugin {
               relName = token.id.replace(match, '');
             }
             setTransform(token.id, {
+              ...query,
               format: FORMAT_TAILWIND,
               localID: makeCSSVar(`${path.join('-')}-${relName.replace(/\./g, '-')}`),
               value: typeof token.value === 'object' ? token.value['.']! : token.value,
-              ...transformContext,
             });
           }
         }
@@ -96,14 +88,12 @@ export default function pluginTailwind(options: TailwindPluginOptions): Plugin {
     },
     async build({ getTransforms, outputFile }) {
       let generatedTheme = '';
-      for (const [variant, input] of Object.entries(variations)) {
-        const transformContext = input && isLegacyModes(input) ? { mode: input.tzMode } : { input };
-
+      for (const [variant, { input, selector }] of Object.entries(variations)) {
         if (generatedTheme) {
           generatedTheme += '\n'; // add extra line break if continuing
         }
-        generatedTheme += `${variant === DEFAULT_THEME ? '@theme' : `@variant ${variant}`} {\n`;
-        for (const token of getTransforms({ format: FORMAT_TAILWIND, ...transformContext })) {
+        generatedTheme += `${variant === DEFAULT_THEME ? '@theme' : `@custom-variant ${variant} (${selector})`} {\n`;
+        for (const token of getTransforms({ ...getTokenQuery(input), format: FORMAT_TAILWIND })) {
           generatedTheme += `  ${token.localID}: ${token.value};\n`;
         }
         generatedTheme += '}\n';
@@ -126,7 +116,12 @@ export default function pluginTailwind(options: TailwindPluginOptions): Plugin {
   };
 }
 
-/** This is a backwards-compat way for someone  */
+/** Convert modes to inputs */
 function isLegacyModes(input: Record<string, string>): boolean {
   return Object.keys(input).length === 0 && 'tzMode' in input;
+}
+
+/** Build query for both resolvers and legacy modes */
+function getTokenQuery(input: Record<string, string>) {
+  return isLegacyModes(input) ? { mode: input.tzMode ?? '.' } : { input };
 }